Concept Art for level 1
I created a sketch of what I would like level one to look like, the
layout is there and it has the basic weapons you will be using
throughout the game. Note how the cage is on the left, it has come to my
attention that it will be on the right, I would still like to discuss
this, but for now that's what the drawing has.